+++ src/main/components/Navigation/FileMenu.tsx
@@ -2,13 +2,13 @@ import { observer } from "mobx-react-lite"
 import { FC } from "react"
 import { useToast } from "use-toast-mui"
 import { localized } from "../../../common/localize/localizedString"
-import { createSong } from "../../actions"
-import { openFile, saveFile, saveFileAs } from "../../actions/file"
-import { useStores } from "../../hooks/useStores"
 import {
   ContextMenuDivider as Divider,
   ContextMenuItem as MenuItem,
-} from "../ContextMenu/ContextMenu"
+} from "../../../components/ContextMenu"
+import { createSong } from "../../actions"
+import { openFile, saveFile, saveFileAs } from "../../actions/file"
+import { useStores } from "../../hooks/useStores"
 
 export const FileMenu: FC<{ close: () => void }> = observer(({ close }) => {
   const rootStore = useStores()
