+++ src/main/components/Navigation/FileMenuButton.tsx
@@ -5,12 +5,12 @@ import Color from "color"
 import { observer } from "mobx-react-lite"
 import { FC, useCallback, useRef } from "react"
 import { localized } from "../../../common/localize/localizedString"
-import { hasFSAccess } from "../../actions/file"
-import { useStores } from "../../hooks/useStores"
 import {
   ContextMenuDivider as Divider,
   ContextMenuItem as MenuItem,
-} from "../ContextMenu/ContextMenu"
+} from "../../../components/ContextMenu"
+import { hasFSAccess } from "../../actions/file"
+import { useStores } from "../../hooks/useStores"
 import { CloudFileMenu } from "./CloudFileMenu"
 import { FileMenu } from "./FileMenu"
 import { LegacyFileMenu } from "./LegacyFileMenu"
