+++ src/main/components/Navigation/LegacyFileMenu.tsx
@@ -2,12 +2,12 @@ import { observer } from "mobx-react-lite"
 import { ChangeEvent, FC } from "react"
 import { useToast } from "use-toast-mui"
 import { localized } from "../../../common/localize/localizedString"
-import { createSong, openSong, saveSong } from "../../actions"
-import { useStores } from "../../hooks/useStores"
 import {
   ContextMenuDivider as Divider,
   ContextMenuItem as MenuItem,
-} from "../ContextMenu/ContextMenu"
+} from "../../../components/ContextMenu"
+import { createSong, openSong, saveSong } from "../../actions"
+import { useStores } from "../../hooks/useStores"
 
 const fileInputID = "OpenButtonInputFile"
