+++ src/main/components/Navigation/UserButtonContent.tsx
@@ -5,8 +5,8 @@ import Color from "color"
 import { User } from "firebase/auth"
 import { FC, useRef, useState } from "react"
 import { localized } from "../../../common/localize/localizedString"
+import { ContextMenuItem as MenuItem } from "../../../components/ContextMenu"
 import { useTheme } from "../../hooks/useTheme"
-import { ContextMenuItem as MenuItem } from "../ContextMenu/ContextMenu"
 import { IconStyle, Tab, TabTitle } from "./Navigation"
 
 const StyledMenu = styled(Menu)`
