+++ src/main/components/TrackList/TrackListContextMenu.tsx
@@ -1,7 +1,7 @@
 import { MenuItem } from "@mui/material"
 import { FC } from "react"
 import { localized } from "../../../common/localize/localizedString"
-import { ContextMenu, ContextMenuProps } from "../ContextMenu/ContextMenu"
+import { ContextMenu, ContextMenuProps } from "../../../components/ContextMenu"
 
 export type TrackListContextMenuProps = ContextMenuProps & {
   onClickAdd: () => void
